+++ src/server.go
@@ -93,6 +93,9 @@ This is <a href="https://github.com/folio-org/mod-reporting">mod-reporting</a>. 
   <li><a href="/ldp/db/tables">List tables from reporting database</a></li>
   <li><a href="/ldp/db/columns?schema=folio_users&table=users">List columns for "users" table</a></li>
   <li><a href="/ldp/db/log">Logs</a></li>
+  <li><a href="/ldp/db/version">Version</a></li>
+  <li><a href="/ldp/db/updates">Updates</a></li>
+  <li><a href="/ldp/db/processes">Processes</a></li>
 </ul>`)
 		return
 	} else if path == "/admin/health" {
@@ -114,6 +117,12 @@ This is <a href="https://github.com/folio-org/mod-reporting">mod-reporting</a>. 
 		runWithErrorHandling(w, req, server, handleReport)
 	} else if path == "/ldp/db/log" {
 		runWithErrorHandling(w, req, server, handleLogs)
+	} else if path == "/ldp/db/version" {
+		runWithErrorHandling(w, req, server, handleVersion)
+	} else if path == "/ldp/db/updates" {
+		runWithErrorHandling(w, req, server, handleUpdates)
+	} else if path == "/ldp/db/processes" {
+		runWithErrorHandling(w, req, server, handleProcesses)
 	} else {
 		// Unrecognized
 		w.WriteHeader(http.StatusNotFound)