+++ src/reporting.go
@@ -364,6 +364,10 @@ func handleLogs(w http.ResponseWriter, req *http.Request, session *ModReportingS
 		return fmt.Errorf("could not find reporting DB: %w", err)
 	}
 
+	if !session.isMDB {
+		return &HTTPError{http.StatusNotImplemented, "Implemented only for MetaDB, not LDP"}
+	}
+
 	rows, err := dbConn.Query(context.Background(), "SELECT log_time, error_severity, message FROM metadb.log")
 	if err != nil {
 		return fmt.Errorf("could not fetch logs from reporting DB: %w", err)
@@ -391,6 +395,10 @@ func handleVersion(w http.ResponseWriter, req *http.Request, session *ModReporti
 	dbConn, err := session.findDbConn(req.Header.Get("X-Okapi-Token"))
 	if err != nil {
 		return fmt.Errorf("could not find reporting DB: %w", err)
+	}
+
+	if !session.isMDB {
+		return &HTTPError{http.StatusNotImplemented, "Implemented only for MetaDB, not LDP"}
 	}
 
 	rows, err := dbConn.Query(context.Background(), "SELECT mdbversion()")
@@ -423,6 +431,10 @@ func handleUpdates(w http.ResponseWriter, req *http.Request, session *ModReporti
 		return fmt.Errorf("could not find reporting DB: %w", err)
 	}
 
+	if !session.isMDB {
+		return &HTTPError{http.StatusNotImplemented, "Implemented only for MetaDB, not LDP"}
+	}
+
 	rows, err := dbConn.Query(context.Background(), "SELECT schema_name, table_name, last_update, elapsed_real_time FROM metadb.table_update ORDER BY elapsed_real_time DESC")
 	if err != nil {
 		return fmt.Errorf("could not fetch updates from reporting DB: %w", err)
@@ -449,6 +461,10 @@ func handleProcesses(w http.ResponseWriter, req *http.Request, session *ModRepor
 	dbConn, err := session.findDbConn(req.Header.Get("X-Okapi-Token"))
 	if err != nil {
 		return fmt.Errorf("could not find reporting DB: %w", err)
+	}
+
+	if !session.isMDB {
+		return &HTTPError{http.StatusNotImplemented, "Implemented only for MetaDB, not LDP"}
 	}
 
 	rows, err := dbConn.Query(context.Background(), "SELECT dbname, username, state, realtime, query FROM ps() ORDER BY realtime DESC")