+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.3.0](https://github.com/folio-org/mod-reporting/tree/v1.3.0) (IN PROGRESS)
 
 * Implement `/ldp/db/log`, add tests. This is in the module descriptor, so it's needed for full compatibility. Fixes MODREP-8.
+* Upgrade dependency on `crypto` library (v0.27.0 has a vulnerability). Fixes MODREP-16.
 
 ## [1.2.0](https://github.com/folio-org/mod-reporting/tree/v1.2.0) (2024-10-29)