+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Implement `/ldp/db/log`, add tests. This is in the module descriptor, so it's needed for full compatibility. Fixes MODREP-8.
 * Upgrade dependency on `crypto` library (v0.27.0 has a vulnerability). Fixes MODREP-16.
 * Add three new WSAPI endpoints (`/ldp/db/version`, `/ldp/db/updates`, `/ldp/db/processes`), write tests, update documentation. Provided interface `ldp-query` bumped from v1.3 to v1.4. Fixes MODREP-2.
-* Each new FOLIO session gets a new database connectio, causing the current DB config to be re-read. Fixes MODREP-11.
+* Each new FOLIO session gets a new reporting-database connection, causing the current DB config to be re-read. Fixes MODREP-11.
 
 ## [1.2.0](https://github.com/folio-org/mod-reporting/tree/v1.2.0) (2024-10-29)